+++ core/common-ui/src/main/java/team/noweekend/core/common/ui/profile/input/BirthInputField.kt
@@ -1,6 +1,5 @@
-package team.noweekend.core.common.ui.profile.inputfield
+package team.noweekend.core.common.ui.profile.input
 
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.text.input.TextFieldState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -13,23 +12,23 @@ import team.noweekend.core.design.system.core.component.input.status.TextInputTy
 import team.noweekend.core.resource.NWKStringResource
 
 @Composable
-internal fun NickNameInputField(
+fun BirthInputField(
     textFieldState: TextFieldState,
     onKeyboardAction: () -> Unit,
-    focusRequester: FocusRequester,
     inputFieldStatus: InputFieldStatus,
     modifier: Modifier = Modifier,
+    focusRequester: FocusRequester = FocusRequester(),
 ) {
     NWKInputField(
-        label = stringResource(NWKStringResource.NicknameInputLabel),
-        placeholder = stringResource(NWKStringResource.NicknameInputPlaceholder),
-        errorText = stringResource(NWKStringResource.NicknameInputErrorMessage),
+        label = stringResource(NWKStringResource.BirthInputLabel),
+        placeholder = stringResource(NWKStringResource.BirthInputPlaceholder),
+        errorText = stringResource(NWKStringResource.BirthInputErrorMessage),
         textFieldState = textFieldState,
         onKeyboardAction = onKeyboardAction,
-        modifier = modifier.fillMaxWidth(),
+        modifier = modifier,
         inputFieldStatus = inputFieldStatus,
         focusRequester = focusRequester,
-        textInputType = TextInputType.NICKNAME,
-        keyboardImeAction = ImeAction.Next,
+        textInputType = TextInputType.DAY,
+        keyboardImeAction = ImeAction.Done,
     )
 }
