+++ core/common-ui/src/main/java/team/noweekend/core/common/ui/profile/input/NickNameInputField.kt
@@ -1,5 +1,6 @@
-package team.noweekend.core.common.ui.profile.inputfield
+package team.noweekend.core.common.ui.profile.input
 
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.text.input.TextFieldState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -12,23 +13,23 @@ import team.noweekend.core.design.system.core.component.input.status.TextInputTy
 import team.noweekend.core.resource.NWKStringResource
 
 @Composable
-fun BirthInputField(
+fun NickNameInputField(
     textFieldState: TextFieldState,
     onKeyboardAction: () -> Unit,
+    focusRequester: FocusRequester,
     inputFieldStatus: InputFieldStatus,
     modifier: Modifier = Modifier,
-    focusRequester: FocusRequester = FocusRequester(),
 ) {
     NWKInputField(
-        label = stringResource(NWKStringResource.BirthInputLabel),
-        placeholder = stringResource(NWKStringResource.BirthInputPlaceholder),
-        errorText = stringResource(NWKStringResource.BirthInputErrorMessage),
+        label = stringResource(NWKStringResource.NicknameInputLabel),
+        placeholder = stringResource(NWKStringResource.NicknameInputPlaceholder),
+        errorText = stringResource(NWKStringResource.NicknameInputErrorMessage),
         textFieldState = textFieldState,
         onKeyboardAction = onKeyboardAction,
-        modifier = modifier,
+        modifier = modifier.fillMaxWidth(),
         inputFieldStatus = inputFieldStatus,
         focusRequester = focusRequester,
-        textInputType = TextInputType.DAY,
-        keyboardImeAction = ImeAction.Done,
+        textInputType = TextInputType.NICKNAME,
+        keyboardImeAction = ImeAction.Next,
     )
 }
